[board] default to SyterBoot for 100ask-d1-h; rename `d1` to `sun20iw1`
