+++ rust/src/soc.rs
@@ -1,1 +1,1 @@
-pub mod d1;
+pub mod sun20iw1;
